Cost Assignment
a) Compare the price of the architecture you created in Assignment 2 for the 3 public clouds. (Make some realistic assumptions if reqd). Submit screenshots of the pricing tools you have used. 
b) Utilize the Total Cost of Ownership Calculator to determine a scenario where moving to cloud is advisable and a scenario where it is not advisable. (The scenarios have to be realistic). Submit screenshots from the TCO tool along with details of the scenarios.